+++ Interview_Prep_App/native_java/retrofit/src/main/res/layout/activity_splash_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/primaryColor"
+    android:background="@color/white"
     tools:context=".ui.main.SplashScreenActivity">
 
     <TextView
