--- Interview_Prep_App/native_kotlin/retrofit/src/main/res/values/splash_theme.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Splash Screen Theme. -->
-    <style name="Theme.AppSplash" parent="Theme.SplashScreen">
-        <item name="windowSplashScreenBackground">@color/white</item>
-        <item name="windowSplashScreenAnimatedIcon">@mipmap/ic_logo_foreground</item>
-        <item name="windowSplashScreenAnimationDuration">2000</item>
-        <item name="postSplashScreenTheme">@style/Theme.AppTheme</item>
-
-        <!-- Status bar and Nav bar configs -->
-        <item name="android:statusBarColor" tools:targetApi="l">@color/white</item>
-        <item name="android:navigationBarColor">@color/white</item>
-<!--        if your minimum API is 26, uncomment the following code to provide lightStatusBar-->
-<!--        <item name="android:windowLightStatusBar">true</item>-->
-    </style>
-</resources>
-<!--<a href="https://www.freepik.com/free-vector/taking-notes-concept-illustration_7069574.htm#query=interview&position=4&from_view=search&track=sph#page=1&query=i&from_query=undefined&position=0&from_view=search&track=sph">Image by storyset</a> on Freepik-->
