+++ Interview_Prep_App/native_kotlin/view_binding/build.gradle
@@ -21,7 +21,7 @@ android {
         jvmTarget = "1.8"
     }
     defaultConfig {
-        applicationId "com.interviewprep.kotlinretrofit"
+        applicationId "com.interviewprep.kotlinviewbinding"
         minSdkVersion 25
         targetSdkVersion 33
         versionCode 1
@@ -37,6 +37,11 @@ android {
     packagingOptions {
         exclude 'META-INF/com.android.tools/proguard/coroutines.pro'
     }
+    // to enable view binding in module, set viewBinding to true in buildFeatures{} when it is enabled
+    // a binding class is generated for each XML layout file.
+    buildFeatures {
+        viewBinding = true
+    }
 }
 
 dependencies {
@@ -46,7 +51,6 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-
     implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.2'
 
@@ -67,4 +71,6 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-moshi:2.9.0"
     implementation 'com.google.android.material:material:1.7.0'
 
+    // Splashscreen
+    implementation "androidx.core:core-splashscreen:1.0.0"
 }
