+++ Interview_Prep_App/native_kotlin/view_binding/src/main/java/com/interviewprep/kotlinviewbinding/repository/models/Answer.kt
@@ -1,0 +1,19 @@
+package com.interviewprep.kotlinviewbinding.repository.models
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+/**
+ * Created by Tc2r on 5/19/2017.
+ *
+ *
+ * Description:
+ */
+@Parcelize
+data class Answer(
+    @SerializedName("answer")
+    var answer: String? = null,
+    @SerializedName("details")
+    var details: String? = null
+) : Parcelable
