+++ Interview_Prep_App/native_kotlin/view_binding/src/main/java/com/interviewprep/kotlinviewbinding/repository/models/response/AnswersResponse.kt
@@ -1,0 +1,9 @@
+package com.interviewprep.kotlinviewbinding.repository.models.response
+
+import com.google.gson.annotations.SerializedName
+import com.interviewprep.kotlinviewbinding.repository.models.Answer
+
+data class AnswersResponse(
+    @SerializedName("answers")
+    var answers: ArrayList<Answer>
+)
