+++ Interview_Prep_App/native_kotlin/view_binding/src/main/java/com/interviewprep/kotlinviewbinding/repository/models/response/QuestionsResponse.kt
@@ -1,0 +1,9 @@
+package com.interviewprep.kotlinviewbinding.repository.models.response
+
+import com.google.gson.annotations.SerializedName
+import com.interviewprep.kotlinviewbinding.repository.models.Question
+
+data class QuestionsResponse(
+    @SerializedName("questions")
+    var questions: ArrayList<Question>
+)
