+++ Interview_Prep_App/native_kotlin/view_binding/src/main/java/com/interviewprep/kotlinviewbinding/util/OnFragmentInteractionListener.kt
@@ -1,0 +1,14 @@
+package com.interviewprep.kotlinviewbinding.util
+
+/**
+ * This interface must be implemented by activities that contain this
+ * fragment to allow an interaction in this fragment to be communicated
+ * to the activity and potentially other fragments contained in that
+ * activity.
+ *
+ *
+ * See the Android Training lesson [Communicating with Other Fragments](http://developer.android.com/training/basics/fragments/communicating.html) for more information.
+ */
+interface OnFragmentInteractionListener {
+    fun fragmentInitialized()
+}
