+++ Interview_Prep_App/native_kotlin/view_binding/src/main/res/layout/activity_main.xml
