--- Interview_Prep_App/native_kotlin/view_binding_/build.gradle
@@ -1,74 +1,0 @@
-plugins {
-    id 'com.android.application'
-    id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
-    id 'kotlin-parcelize'
-//    id 'androidx.navigation.safeargs.kotlin'
-}
-
-android {
-    compileSdkVersion 33
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-    composeOptions {
-        kotlinCompilerExtensionVersion compose_version
-    }
-
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-    defaultConfig {
-        applicationId "com.interviewprep.kotlinretrofit"
-        minSdkVersion 25
-        targetSdkVersion 33
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-    packagingOptions {
-        exclude 'META-INF/com.android.tools/proguard/coroutines.pro'
-    }
-    // to enable view binding in module, set viewBinding to true in buildFeatures{} when it is enabled
-    // a binding class is generated for each XML layout file.
-    buildFeatures {
-        viewBinding = true
-    }
-}
-
-dependencies {
-    implementation project (':sharedresources')
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.9.0'
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-
-    implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.2'
-
-    //Retrofit & Scalar Converter
-    implementation "com.squareup.retrofit2:retrofit:2.9.0"
-    implementation "com.squareup.retrofit2:converter-scalars:2.9.0"
-
-    //GSON
-    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
-
-
-    // Coroutines
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.2"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.2"
-    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.5.1"
-
-    //Retrofit with Moshi converter
-    implementation "com.squareup.retrofit2:converter-moshi:2.9.0"
-    implementation 'com.google.android.material:material:1.7.0'
-}
