--- Interview_Prep_App/native_kotlin/view_binding_/src/main/java/com/interviewprep/kotlinretrofit/repository/models/Answer.kt
@@ -1,19 +1,0 @@
-package com.interviewprep.kotlinretrofit.repository.models
-
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.parcelize.Parcelize
-
-/**
- * Created by Tc2r on 5/19/2017.
- *
- *
- * Description:
- */
-@Parcelize
-data class Answer(
-    @SerializedName("answer")
-    var answer: String? = null,
-    @SerializedName("details")
-    var details: String? = null
-) : Parcelable
