--- Interview_Prep_App/native_kotlin/view_binding_/src/main/java/com/interviewprep/kotlinretrofit/repository/models/response/QuestionsResponse.kt
@@ -1,9 +1,0 @@
-package com.interviewprep.kotlinretrofit.repository.models.response
-
-import com.google.gson.annotations.SerializedName
-import com.interviewprep.kotlinretrofit.repository.models.Question
-
-data class QuestionsResponse(
-    @SerializedName("questions")
-    var questions: ArrayList<Question>
-)
