--- Interview_Prep_App/native_kotlin/view_binding_/src/main/java/com/interviewprep/kotlinretrofit/ui/splash_screen/SplashScreenActivity.kt
@@ -1,24 +1,0 @@
-package com.interviewprep.kotlinretrofit.ui.splash_screen
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import com.interviewprep.kotlinretrofit.R
-import com.interviewprep.kotlinretrofit.ui.main.MainActivity
-
-class SplashScreenActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash_screen)
-
-        supportActionBar?.hide()
-
-        Handler(Looper.getMainLooper()).postDelayed({
-            val intent = Intent(this@SplashScreenActivity, MainActivity::class.java)
-            startActivity(intent)
-            finish()
-        }, 2000)
-    }
-}
