--- Interview_Prep_App/native_kotlin/view_binding_/src/main/res/layout/activity_splash_screen.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.splash_screen.SplashScreenActivity"
-    android:background= "#fffde7">
-
-    <ImageView
-        android:id="@+id/imgSplash"
-        android:layout_width="200dp"
-        android:layout_height="200dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="240dp"
-        android:src="@drawable/splash_screen_logo" />
-
-    <TextView
-        android:id="@+id/txtSplash"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/imgSplash"
-        android:layout_centerHorizontal="true"
-        android:text="@string/app_name"
-        android:textColor="@color/black"
-        android:textStyle="bold"
-        android:textSize="29sp"/>
-
-</RelativeLayout>
