+++ Interview_Prep_App/settings.gradle
@@ -31,4 +31,4 @@ include ':sharedresources'
 include ':native_kotlin:template'
 include ':native_java:javatemplate'
 include ':native_kotlin:clean-architecture'
-include ':native_kotlin:view_binding_'
+include ':native_kotlin:view_binding'
